Remove quotes from CSV data
Espoo BI tooling doesn't follow common CSV standards (e.g. RFC4180) and
simply can't handle quote characters escaped as double quotes
